First Embedded Push: Prototype 2
This Prototype code will test a desired number of pins and has a preset
value in which pins will pass or fail. This value will need to be
variable once the  GUI code is combined with the Embedded code. This
code also prints to PORTC pins 1 and 0 a pass or fail for each pin.